// src/models/IoK8sApiAutoscalingV2MetricIdentifier/types.ts
import { IoK8sApimachineryPkgApisMetaV1LabelSelectorModelOptions as IoK8sApimachineryPkgApisMetaV1LabelSelector } from '../IoK8sApimachineryPkgApisMetaV1LabelSelector';

export interface IoK8sApiAutoscalingV2MetricIdentifierModelOptions {
  name: string;
  selector?: IoK8sApimachineryPkgApisMetaV1LabelSelector;
}